code cleanup: merge dso_api.lib.exceptions with rest_framework_dso.exceptions
